requirements.txt
asgiref==3.7.2
distlib==0.3.7
Django==4.2.5
django-rest-framework==0.1.0
djangorestframework==3.14.0
filelock==3.12.4
Pillow==10.0.1
platformdirs==3.11.0
pytz==2023.3.post1
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.5